backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets>=11.0,<13.0
httpx>=0.24,<0.26
python-dotenv==1.0.1
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
numpy==1.24.3
# Authentication dependencies
bcrypt==4.1.2
PyJWT==2.8.0
supabase==2.3.4
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.19
databases[asyncpg]==0.7.3
asyncpg==0.27.0
alembic==1.11.1
psycopg2-binary==2.9.7